+++ app/models/project.rb
@@ -74,6 +74,7 @@ class Project < ApplicationRecord
   has_many :treatment_arms, -> { current.joins(:randomization_scheme).merge(RandomizationScheme.current) }
   has_many :grid_variables
   has_many :invites
+  has_many :project_languages
 
   # Methods